simplify loadbalancer creation
We´ll need to create more loadbalancers for idling and service
affinity.

Current code creates the loadbalancers one by one, that doesn´t
scale well and is ugly to read.

We can simplify the loadbalancer creation, and also remove the
loadbalancer uuid variables from the controller, that only are
used once, and is not expensive to ask directly to the nbctl
cache.

Signed-off-by: Antonio Ojea <aojea@redhat.com>